@@ -1,4 +1,4 @@
-import { FormGroup, NgForm } from "@angular/forms";
+import { NgForm, FormGroup } from "@angular/forms";
 
 export interface Form {
     formGroup: FormGroup;
